Move error handling test to an integration test
